Hello world, hello reader.
This is my first commit on this project, which is hopefully going to be
a nice wysiwyg editor component for Cycle.js. I've started out by
creating a skeleton for my app and rendering Hello World to the DOM.

The skeleton is basically identical to skeletons in my other projects,
but I actually wrote this one out by hand. I think it's good to be able
to write this out from scratch.